% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyCharts_extended.R
\name{aeExplorerOutput}
\alias{aeExplorerOutput}
\title{Shiny bindings for safetyCharts aeExplorer widget}
\usage{
aeExplorerOutput(id)
}
\arguments{
\item{id}{\code{character} The id for this element.}
}
\value{
An \code{\link[htmlwidgets:htmlwidgets-shiny]{htmlwidgets::shinyWidgetOutput()}} output function.
}
\description{
Shiny bindings for safetyCharts aeExplorer widget
}
\keyword{internal}
