% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyCharts_extended.R
\name{aeTimelinesOutput}
\alias{aeTimelinesOutput}
\title{Shiny bindings for safetyCharts aeTimelines widget}
\usage{
aeTimelinesOutput(id)
}
\arguments{
\item{id}{\code{character} The id for this element.}
}
\value{
An \code{\link[htmlwidgets:htmlwidgets-shiny]{htmlwidgets::shinyWidgetOutput()}} output function.
}
\description{
Shiny bindings for safetyCharts aeTimelines widget
}
\keyword{internal}
