% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_AE.R
\name{mod_AE_Server}
\alias{mod_AE_Server}
\title{Adverse Events Plugin Server}
\usage{
mod_AE_Server(id, rctv_dfAE, rctv_dfSUBJ, rctv_strSiteID, rctv_strSubjectID)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{rctv_dfAE}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns an adverse events domain dataframe.}

\item{rctv_dfSUBJ}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a subject domain dataframe.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected site, and can be used to update
the selected site.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}
}
\value{
A module server function.
}
\description{
Adverse Events Plugin Server
}
