% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_AE.R
\name{mod_AE_UI}
\alias{mod_AE_UI}
\title{Adverse Events Plugin UI}
\usage{
mod_AE_UI(id)
}
\arguments{
\item{id}{\code{character} The id for this element.}
}
\value{
A \code{\link[bslib:page_fillable]{bslib::page_fillable()}} with the AE plugin UI.
}
\description{
Adverse Events Plugin UI
}
