% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_AEChartsTitle.R
\name{mod_AEChartsTitle_Server_Color}
\alias{mod_AEChartsTitle_Server_Color}
\title{Charts Panel Title Block Server (by color)}
\usage{
mod_AEChartsTitle_Server_Color(
  id,
  chrColors = c("#1b9e77", "#d95f02", "#7570b3"),
  rctv_strGroupID,
  rctv_strGroupLevel,
  rctv_strSubjectID
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{chrColors}{\code{character} Colors to use for a scale, labels, etc.}

\item{rctv_strGroupID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected group (usually site), and can be
used to update which group is selected.}

\item{rctv_strGroupLevel}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}}
object that returns the selected \code{GroupLevel}.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} with the selected category.
}
\description{
Charts Panel Title Block Server (by color)
}
\keyword{internal}
