% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_AEChartsTitle.R
\name{mod_AEChartsTitle_UI}
\alias{mod_AEChartsTitle_UI}
\title{Charts Panel Title Block UI}
\usage{
mod_AEChartsTitle_UI(id, chrFields)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{chrFields}{\code{character} A named vector of fields in \code{dfAE} to display in
a drop-down menu, where the names are display values and the values are
column names. Passed to \code{\link[shinyWidgets:virtualSelectInput]{shinyWidgets::virtualSelectInput()}} as \code{choices}.}
}
\value{
A \code{\link[shiny:reexports]{shiny::div()}} with text "Prevalence of", then the category
dropdown, then a span that will be filled with information about the
grouping.
}
\description{
Charts Panel Title Block UI
}
\keyword{internal}
