% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_AEDashboard.R
\name{mod_AEDashboard_UI}
\alias{mod_AEDashboard_UI}
\title{Dashboard Tab UI}
\usage{
mod_AEDashboard_UI(id)
}
\arguments{
\item{id}{\code{character} The id for this element.}
}
\value{
A \code{\link[bslib:layout_columns]{bslib::layout_columns()}} layout with information about the AE
domain.
}
\description{
Dashboard Tab UI
}
\keyword{internal}
