% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_AESummary.R
\name{mod_AESummary_Server}
\alias{mod_AESummary_Server}
\title{Adverse Events Summary Card Server}
\usage{
mod_AESummary_Server(id, rctv_dateSnapshot, rctv_dfAE, rctv_dfSUBJ)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{rctv_dateSnapshot}{\verb{reactive Date} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the date of a data snapshot.}

\item{rctv_dfAE}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns an adverse events domain dataframe.}

\item{rctv_dfSUBJ}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a subject domain dataframe.}
}
\value{
A module server function.
}
\description{
Adverse Events Summary Card Server
}
\keyword{internal}
