% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_AESummary.R
\name{mod_AESummary_UI}
\alias{mod_AESummary_UI}
\title{Adverse Events Summary Card UI}
\usage{
mod_AESummary_UI(id)
}
\arguments{
\item{id}{\code{character} The id for this element.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with counts of AEs, SAEs, participants, and
participants without AEs.
}
\description{
Adverse Events Summary Card UI
}
\keyword{internal}
