% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_AETimeline.R
\name{mod_AETimeline_Server}
\alias{mod_AETimeline_Server}
\title{Adverse Events Timeline Panel Server}
\usage{
mod_AETimeline_Server(
  id,
  rctv_dfAE_Study,
  rctv_strGroupID,
  rctv_strGroupLevel,
  rctv_strSubjectID,
  chrDateFields = c(mincreated_dts = "AE Created", aest_dt = "AE Start", aeen_dt =
    "AE End")
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{rctv_dfAE_Study}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a study-level adverse events domain dataframe.}

\item{rctv_strGroupID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected group (usually site), and can be
used to update which group is selected.}

\item{rctv_strGroupLevel}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}}
object that returns the selected \code{GroupLevel}.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}

\item{chrDateFields}{\code{character} A named vector of fields in \code{dfAE} to treat
as dates, where the names are the column names and the values are what to
display to users. Unnamed elements will be used as both the field name and
the display name. These fields are currently used in the Timeline panel.}
}
\value{
A module server function.
}
\description{
Adverse Events Timeline Panel Server
}
\keyword{internal}
