% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_AETimeline.R
\name{mod_AETimeline_UI}
\alias{mod_AETimeline_UI}
\title{Adverse Events Timeline Panel UI}
\usage{
mod_AETimeline_UI(id)
}
\arguments{
\item{id}{\code{character} The id for this element.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} for visualizing date variables in the AE data.
}
\description{
Adverse Events Timeline Panel UI
}
\keyword{internal}
