% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_AEPrevalence.R
\name{out_AEPrevalence_Expanded}
\alias{out_AEPrevalence_Expanded}
\title{UI for Expanded Prevalence Panel}
\usage{
out_AEPrevalence_Expanded(ns)
}
\arguments{
\item{ns}{\code{function} A namespace function, such as the one returned by
\code{\link[shiny:NS]{shiny::NS()}}.}
}
\value{
A shiny tag object, usually for use in \code{\link[shiny:renderUI]{shiny::renderUI()}}.
}
\description{
UI for Expanded Prevalence Panel
}
\keyword{internal}
