% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_AEDashboard.R
\name{out_DashboardCard}
\alias{out_DashboardCard}
\title{Standardized cards for dashboard tab}
\usage{
out_DashboardCard(id, title, ...)
}
\arguments{
\item{id}{\code{character} The id for this element.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with appropriate size and title.
}
\description{
Standardized cards for dashboard tab
}
\keyword{internal}
