% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluginAE.R
\name{pluginAE}
\alias{pluginAE}
\title{AE Plugin Definition}
\usage{
pluginAE(
  chrCategoricalFields = c(aeser = "Serious?", mdrpt_nsv = "Preferred Term", mdrsoc_nsv =
    "System Organ Class", aetoxgr = "Toxicity Grade", aeongo = "Ongoing?", aerel =
    "Related?"),
  chrDateFields = c(mincreated_dts = "AE Entry Date", aest_dt = "AE Start Date", aeen_dt
    = "AE End Date"),
  strMetricID_AE = "Analysis_kri0001",
  strMetricID_SAE = "Analysis_kri0002"
)
}
\arguments{
\item{chrCategoricalFields}{\code{character} A named vector of fields in \code{dfAE} to
treat as categorical, where the names are the column names and the values
are what to display to users. Unnamed elements will be used as both the
field name and the display name. These fields are currently used in the
Prevalence panel.}

\item{chrDateFields}{\code{character} A named vector of fields in \code{dfAE} to treat
as dates, where the names are the column names and the values are what to
display to users. Unnamed elements will be used as both the field name and
the display name. These fields are currently used in the Timeline panel.}

\item{strMetricID_AE}{\verb{length-1 character} The ID of the adverse events
metric. Defaults to \code{"Analysis_kri0001"}.}

\item{strMetricID_SAE}{\verb{length-1 character} The ID of the serious adverse
events metric. Defaults to \code{"Analysis_kri0001"}.}
}
\value{
A plugin definition for use in gsm.app.
}
\description{
Plugins for gsm.app are defined by a named \code{list} with information about the
plugin. These can be loaded from a yaml file with \code{\link[gsm.app:plugin_Read]{gsm.app::plugin_Read()}}.
We provide a ready-to-use plugin definition for this AE plugin.
}
\examples{
pluginAE()
}
