% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyCharts_extended.R
\name{render_aeExplorer}
\alias{render_aeExplorer}
\title{Shiny bindings for safetyCharts aeExplorer widget}
\usage{
render_aeExplorer(expr, envEvaluate = parent.frame())
}
\arguments{
\item{envEvaluate}{\code{environment} The environment in which any variables used
in the message or widget are defined. You almost definitely want to leave
this as the default value.}
}
\value{
An \code{\link[htmlwidgets:htmlwidgets-shiny]{htmlwidgets::shinyRenderWidget()}} render function.
}
\description{
Shiny bindings for safetyCharts aeExplorer widget
}
\keyword{internal}
