% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyCharts_extended.R
\name{render_SC_Widget}
\alias{render_SC_Widget}
\title{Prepare and render a safetyCharts widget}
\usage{
render_SC_Widget(lData, lSettings, fnInit, strWidgetName)
}
\arguments{
\item{lData}{\code{list} Raw data for use in a safetyCharts widget.}

\item{lSettings}{\code{list} Named list of settings for a safetyCharts widget.}

\item{fnInit}{\code{function} An initialization function such as
\code{\link[safetyCharts:init_aeExplorer]{safetyCharts::init_aeExplorer()}}.}

\item{strWidgetName}{\code{character} The name of a widget in the gsm package.}
}
\value{
The result of a call to \code{\link[safetyCharts:render_widget]{safetyCharts::render_widget()}} with the
initialized data.
}
\description{
Prepare and render a safetyCharts widget
}
\keyword{internal}
