% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{scale_fill_StudyGroupSubject}
\alias{scale_fill_StudyGroupSubject}
\title{Fill Scale for Study, Group, and Subject}
\usage{
scale_fill_StudyGroupSubject(
  chrColors = c(Study = "#1b9e77", Group = "#d95f02", Subject = "#7570b3")
)
}
\arguments{
\item{chrColors}{\code{character} Colors to use for a scale, labels, etc.}
}
\value{
A \code{\link[ggplot2:scale_colour_discrete]{ggplot2::scale_fill_discrete()}}.
}
\description{
Fill Scale for Study, Group, and Subject
}
\keyword{internal}
