% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyCharts_extended.R
\name{SCWidgetOutput}
\alias{SCWidgetOutput}
\title{Shiny bindings for safetyCharts widgets}
\usage{
SCWidgetOutput(id, strWidgetName)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{strWidgetName}{\code{character} The name of a widget in the gsm package.}
}
\value{
An \code{\link[htmlwidgets:htmlwidgets-shiny]{htmlwidgets::shinyWidgetOutput()}} output function.
}
\description{
Shiny bindings for safetyCharts widgets
}
\keyword{internal}
