% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared.R
\name{shared-params}
\alias{shared-params}
\title{Parameters used in multiple functions}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{rctv_dateSnapshot}{\verb{reactive Date} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the date of a data snapshot.}

\item{rctv_dfAE}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns an adverse events domain dataframe.}

\item{rctv_dfSUBJ}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a subject domain dataframe.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected site, and can be used to update
the selected site.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}

\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}
}
\description{
Reused parameter definitions are gathered here for easier usage. Note: We
intentionally use \code{id} rather than the more standardized \code{strID} throughout
this package, because \code{\link[shiny:testServer]{shiny::testServer()}} specifically expects an \code{id}
argument in module server functions.
}
\keyword{internal}
