% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyCharts_extended.R
\name{shinyRenderSCWidget}
\alias{shinyRenderSCWidget}
\title{Shiny bindings for safetyCharts widgets}
\usage{
shinyRenderSCWidget(expr, envEvaluate, fnWidgetOutput)
}
\arguments{
\item{envEvaluate}{\code{environment} The environment in which any variables used
in the message or widget are defined. You almost definitely want to leave
this as the default value.}

\item{fnWidgetOutput}{\code{function} An \code{\link[htmlwidgets:htmlwidgets-shiny]{htmlwidgets::shinyWidgetOutput()}}
function.}
}
\value{
An \code{\link[htmlwidgets:htmlwidgets-shiny]{htmlwidgets::shinyRenderWidget()}} render function.
}
\description{
Shiny bindings for safetyCharts widgets
}
\keyword{internal}
