% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-names.R
\name{SwapNamesForValues}
\alias{SwapNamesForValues}
\title{Swap Names and Values of an Object}
\usage{
SwapNamesForValues(x)
}
\arguments{
\item{x}{\code{character} A character vector, or something that can be coerced to
one.}
}
\value{
A character vector, with what were previously the names as the
values, and what were previously the values as names.
}
\description{
Swap Names and Values of an Object
}
\keyword{internal}
