% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-names.R
\name{ValidateNames}
\alias{ValidateNames}
\title{Make Sure a Vector of Fields is Named}
\usage{
ValidateNames(chrFields, envCall = rlang::caller_env())
}
\arguments{
\item{chrFields}{\code{character} A named vector of fields in \code{dfAE} to display in
a drop-down menu, where the names are display values and the values are
column names. Passed to \code{\link[shinyWidgets:virtualSelectInput]{shinyWidgets::virtualSelectInput()}} as \code{choices}.}

\item{envCall}{\code{environment} The environment from which this function was
called, for use in better error messages. This value should usually be left
as the default, or passed from the calling function if the calling function
also has an \code{envCall} argument.}
}
\value{
The vector, named.
}
\description{
Make Sure a Vector of Fields is Named
}
\keyword{internal}
