% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_AE.R
\name{mod_AE_Server}
\alias{mod_AE_Server}
\title{Adverse Events Plugin Server}
\usage{
mod_AE_Server(
  id,
  dfAnalyticsInput,
  dfResults,
  rctv_dSnapshotDate,
  rctv_dfAE_Study,
  rctv_strGroupID,
  rctv_strGroupLevel,
  rctv_strSubjectID,
  chrCategoricalFields = c(aeser = "Serious?", mdrpt_nsv = "Preferred Term", mdrsoc_nsv =
    "System Organ Class", aetoxgr = "Toxicity Grade", aeongo = "Ongoing?", aerel =
    "Related?"),
  chrDateFields = c(mincreated_dts = "AE Entry Date", aest_dt = "AE Start Date", aeen_dt
    = "AE End Date"),
  strMetricID_AE = "Analysis_kri0001",
  strMetricID_SAE = "Analysis_kri0002"
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.}

\item{rctv_dSnapshotDate}{\verb{reactive Date} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the date of a data snapshot.}

\item{rctv_dfAE_Study}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a study-level adverse events domain dataframe.}

\item{rctv_strGroupID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected group (usually site), and can be
used to update which group is selected.}

\item{rctv_strGroupLevel}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}}
object that returns the selected \code{GroupLevel}.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}

\item{chrCategoricalFields}{\code{character} A named vector of fields in \code{dfAE} to
treat as categorical, where the names are the column names and the values
are what to display to users. Unnamed elements will be used as both the
field name and the display name. These fields are currently used in the
Prevalence panel.}

\item{chrDateFields}{\code{character} A named vector of fields in \code{dfAE} to treat
as dates, where the names are the column names and the values are what to
display to users. Unnamed elements will be used as both the field name and
the display name. These fields are currently used in the Timeline panel.}

\item{strMetricID_AE}{\verb{length-1 character} The ID of the adverse events
metric. Defaults to \code{"Analysis_kri0001"}.}

\item{strMetricID_SAE}{\verb{length-1 character} The ID of the serious adverse
events metric. Defaults to \code{"Analysis_kri0001"}.}
}
\value{
A module server function.
}
\description{
Adverse Events Plugin Server
}
