% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_AEDashboard.R
\name{mod_AEDashboard_Server}
\alias{mod_AEDashboard_Server}
\title{Dashboard Tab Server}
\usage{
mod_AEDashboard_Server(
  id,
  dfAnalyticsInput,
  rctv_dSnapshotDate,
  rctv_dSnapshotDatePrevious,
  rctv_dfAE_Study,
  rctv_strGroupID,
  rctv_strGroupLevel,
  rctv_strSubjectID
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}

\item{rctv_dSnapshotDate}{\verb{reactive Date} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the date of a data snapshot.}

\item{rctv_dSnapshotDatePrevious}{\verb{reactive Date} A \code{\link[shiny:reactive]{shiny::reactive()}}
object that returns the date of a data snapshot immediately before
\code{rctv_dSnapshotDate}, or \code{NA} if there aren't any snapshots before the
selected date.}

\item{rctv_dfAE_Study}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a study-level adverse events domain dataframe.}

\item{rctv_strGroupID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected group (usually site), and can be
used to update which group is selected.}

\item{rctv_strGroupLevel}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}}
object that returns the selected \code{GroupLevel}.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}
}
\value{
A module server function.
}
\description{
Dashboard Tab Server
}
\keyword{internal}
