% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_AESummary.R
\name{mod_AESummary_Server}
\alias{mod_AESummary_Server}
\title{Adverse Events Summary Card Server}
\usage{
mod_AESummary_Server(
  id,
  dfAnalyticsInput,
  dfResults,
  rctv_dSnapshotDate,
  rctv_dSnapshotDatePrevious,
  rctv_strSiteID,
  strMetricID_AE = "Analysis_kri0001",
  strMetricID_SAE = "Analysis_kri0002"
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.}

\item{rctv_dSnapshotDate}{\verb{reactive Date} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the date of a data snapshot.}

\item{rctv_dSnapshotDatePrevious}{\verb{reactive Date} A \code{\link[shiny:reactive]{shiny::reactive()}}
object that returns the date of a data snapshot immediately before
\code{rctv_dSnapshotDate}, or \code{NA} if there aren't any snapshots before the
selected date.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected site, and can be used to update
the selected site.}

\item{strMetricID_AE}{\verb{length-1 character} The ID of the adverse events
metric. Defaults to \code{"Analysis_kri0001"}.}

\item{strMetricID_SAE}{\verb{length-1 character} The ID of the serious adverse
events metric. Defaults to \code{"Analysis_kri0001"}.}
}
\value{
A module server function.
}
\description{
Adverse Events Summary Card Server
}
\keyword{internal}
