% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluginAE.R
\name{pluginAE}
\alias{pluginAE}
\title{AE Plugin Definition}
\usage{
pluginAE(
  chrCategoricalFields = c(aeser = "Serious?", mdrpt_nsv = "Preferred Term", mdrsoc_nsv =
    "System Organ Class", aetoxgr = "Toxicity Grade", aeongo = "Ongoing?", aerel =
    "Related?"),
  chrDateFields = c(mincreated_dts = "AE Entry Date", aest_dt = "AE Start Date", aeen_dt
    = "AE End Date"),
  chrMetricID_AE = c(Site = "Analysis_kri0001", Country = "Analysis_cou0001"),
  chrMetricID_SAE = c(Site = "Analysis_kri0002", Country = "Analysis_cou0002"),
  lWorkflows = list()
)
}
\arguments{
\item{chrCategoricalFields}{\code{character} A named vector of fields in \code{dfAE} to
treat as categorical, where the names are the column names and the values
are what to display to users. Unnamed elements will be used as both the
field name and the display name. These fields are currently used in the
Prevalence panel.}

\item{chrDateFields}{\code{character} A named vector of fields in \code{dfAE} to treat
as dates, where the names are the column names and the values are what to
display to users. Unnamed elements will be used as both the field name and
the display name. These fields are currently used in the Timeline panel.}

\item{chrMetricID_AE}{\code{character} The ID(s) of the adverse events metric(c).
The default value is named by \code{GroupLevel}, but the names are not currently
used. It is advisable to only provide one \code{MetricID} per \code{GroupLevel},
though.}

\item{chrMetricID_SAE}{\code{character} The ID(s) of the serious adverse events
metric(s). The default value is named by \code{GroupLevel}, but the names are
not currently used. It is advisable to only provide one \code{MetricID} per
\code{GroupLevel}, though.}

\item{lWorkflows}{\code{list} An optional named list of workflows to run in order
to translate app domain data into the format required by the plugin (as
defined in the plugin \code{spec}). These workflows are ran before the plugin's
\code{spec} is applied, and are meant to serve as a bridge between your domain
data and the plugin's expected domains.}
}
\value{
A plugin definition for use in gsm.app.
}
\description{
Plugins for gsm.app are defined by a named \code{list} with information about the
plugin. We provide a ready-to-use plugin definition for this AE plugin.
}
\examples{
pluginAE()
}
