% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared.R
\name{shared-params}
\alias{shared-params}
\title{Parameters used in multiple functions}
\arguments{
\item{envCall}{\code{environment} The environment from which this function was
called, for use in better error messages. This value should usually be left
as the default, or passed from the calling function if the calling function
also has an \code{envCall} argument.}

\item{envEvaluate}{\code{environment} The environment in which any variables used
in the message or widget are defined. You almost definitely want to leave
this as the default value.}

\item{fnInit}{\code{function} An initialization function such as
\code{\link[safetyCharts:init_aeExplorer]{safetyCharts::init_aeExplorer()}}.}

\item{fnWidgetOutput}{\code{function} An \code{\link[htmlwidgets:htmlwidgets-shiny]{htmlwidgets::shinyWidgetOutput()}}
function.}

\item{id}{\code{character} The id for this element.}

\item{lData}{\code{list} Raw data for use in a safetyCharts widget.}

\item{lSettings}{\code{list} Named list of settings for a safetyCharts widget.}

\item{rctv_dfAE}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns an adverse events domain dataframe.}

\item{rctv_dfSUBJ}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a subject domain dataframe.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected site, and can be used to update
the selected site.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}

\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}

\item{strWidgetName}{\code{character} The name of a widget in the gsm package.}
}
\description{
Reused parameter definitions are gathered here for easier usage. Note: We
intentionally use \code{id} rather than the more standardized \code{strID} throughout
this package, because \code{\link[shiny:testServer]{shiny::testServer()}} specifically expects an \code{id}
argument in module server functions.
}
\keyword{internal}
